// src/types.ts
export type AccoutrementInput = Omit<Accoutrement, 'id' | 'selected'>;
export type ActionResult = { success: boolean; message?: string; error?: string; service?: Service };

export interface Accoutrement {
  id: string;
  name: string;
  price: number; // in Toman
  durationMinutes: number;
  description: string;
  isComplimentary?: boolean;
  selected?: boolean;
}